Keep window switching off the app scene graph
Switching between two session windows still stalled 100-200ms per switch after #100. Profiles of the release app showed the remaining cost was not inventory work but SwiftUI scene invalidation: GhosthubApp read three @FocusedValue properties at the App level, so every key-window change dirtied the whole App scene. SwiftUI then rebuilt the menu bar and re-ran the WindowGroup content for every open window, twice per switch (synchronously inside the mouse-down and again at layout).

Menu commands now live in dedicated Commands types that read their own focused values, so a focus change invalidates only the menu graph and window switching no longer re-runs each window's root body. Menu behavior and shortcut resolution are unchanged; the helpers moved verbatim.

An A/B profile of otherwise-identical debug builds under the same two-window switching workload cut main-thread busy time from 29% to 14% and removed 99% of the scene rebuilds (scenesDidChange 2814 -> 40 samples). The residual is SwiftUI's inherent active/inactive re-render, about half of which is repeated WorkspaceSidebarModel.sections computation, tracked separately.

Validation: sampled release 0.8.0 and debug builds under scripted 20s window-switching workloads (baseline, focused-value ablation, final fix) and compared main-thread profiles.
