+++ Sources/App/App.swift
@@ -84,11 +84,6 @@ struct GhosthubApp: App {
     private let updateRelaunchStore = UpdateRelaunchManifestStore()
     private let updateRelaunchRestorer = UpdateRelaunchRestorer()
     #endif
-    @FocusedValue(\.sceneModel) private var focusedSceneModel
-    @FocusedValue(\.availableSiblingShortcuts)
-    private var availableSiblingShortcuts
-    @FocusedValue(\.terminalHasEffectiveKeyboardFocus)
-    private var terminalHasEffectiveKeyboardFocus
     @Environment(\.openWindow) private var openWindow
     @State private var didRequestLaunchActivation = false
     #if canImport(AppKit)
@@ -206,32 +201,12 @@ struct GhosthubApp: App {
                 }
                 .keyboardShortcut("q")
             }
-            CommandGroup(after: .appInfo) {
-                Button("Check for Updates…") {
-                    updateController.checkForUpdates()
-                }
-                .disabled(!updateController.isAvailable)
-                Divider()
-                Button("Settings...") {
-                    focusedSceneModel?
-                        .isSettingsPresented = true
-                }
-                .keyboardShortcut(",")
-                Button("Reload Configuration") {
-                    invoke(.reloadConfiguration)
-                }
-                .keyboardShortcut(shortcut(.reloadConfiguration))
-                Divider()
-                Button("Application Log") {
-                    invoke(.openApplicationLog)
-                }
-                .keyboardShortcut(shortcut(.openApplicationLog))
-            }
+            AppMenuCommands(updateController: updateController)
             CommandGroup(replacing: .toolbar) {}
             editMenuCommands
-            fileMenuCommands
-            sessionMenuCommands
-            viewMenuCommands
+            FileMenuCommands(applicationDelegate: appDelegate)
+            SessionMenuCommands()
+            ViewMenuCommands()
             CommandGroup(after: .windowArrangement) {
                 Button("Previous Tab") {
                     NativeTabCommands.selectPrevious()
@@ -324,261 +299,4 @@ struct GhosthubApp: App {
         }
     }
 
-    // MARK: - File menu
-
-    @CommandsBuilder
-    private var fileMenuCommands: some Commands {
-        CommandGroup(replacing: .newItem) {
-            Button("New Worktree…") {
-                invoke(.newWorktree)
-            }
-            .keyboardShortcut(shortcut(
-                .newWorktree,
-                actionIsAvailable:
-                focusedSceneModel?.canCreateWorktreeInSelectedProject == true
-            ))
-            .disabled(
-                focusedSceneModel?.canCreateWorktreeInSelectedProject
-                    != true
-            )
-
-            Button("Import Pull Request…") {
-                invoke(.importPullRequest)
-            }
-            .keyboardShortcut(shortcut(
-                .importPullRequest,
-                actionIsAvailable:
-                focusedSceneModel?.canImportPullRequestInSelectedProject
-                    == true
-            ))
-            .disabled(
-                focusedSceneModel?.canImportPullRequestInSelectedProject
-                    != true
-            )
-
-            Divider()
-
-            Button("New Window") {
-                appDelegate.requestNewWorkspaceWindow()
-            }
-            .keyboardShortcut("n")
-
-            Button("New Tab") {
-                appDelegate.requestNewWorkspaceTab()
-            }
-            .keyboardShortcut("t")
-
-            Divider()
-
-            Button("Split Right") {
-                invoke(.splitRight)
-            }
-            .keyboardShortcut(splitShortcut(.splitRight))
-            .disabled(focusedSceneModel?.canSplitActivePane != true)
-
-            Button("Split Down") {
-                invoke(.splitDown)
-            }
-            .keyboardShortcut(splitShortcut(.splitDown))
-            .disabled(focusedSceneModel?.canSplitActivePane != true)
-
-            Divider()
-
-            Button("Close") {
-                NotificationCenter.default.post(
-                    name: .ghosthubCloseTab,
-                    object: nil
-                )
-            }
-            .keyboardShortcut("w")
-
-            Button("Close Window") {
-                appDelegate.requestWorkspaceWindowClose(
-                    NSApplication.shared.keyWindow
-                )
-            }
-            .keyboardShortcut(
-                "w",
-                modifiers: [.command, .shift]
-            )
-        }
-    }
-
-    // MARK: - Session menu
-
-    @CommandsBuilder
-    private var sessionMenuCommands: some Commands {
-        CommandMenu("Session") {
-            Button("Previous Sibling") {
-                invoke(.previousSibling)
-            }
-            .keyboardShortcut(shortcut(
-                .previousSibling,
-                actionIsAvailable: availableSiblingShortcuts?
-                    .contains(.previousSibling) == true
-            ))
-            .disabled(
-                availableSiblingShortcuts?.contains(.previousSibling) != true
-            )
-
-            Button("Next Sibling") {
-                invoke(.nextSibling)
-            }
-            .keyboardShortcut(shortcut(
-                .nextSibling,
-                actionIsAvailable: availableSiblingShortcuts?
-                    .contains(.nextSibling) == true
-            ))
-            .disabled(
-                availableSiblingShortcuts?.contains(.nextSibling) != true
-            )
-
-            Menu("Select Sibling") {
-                ForEach(
-                    ApplicationShortcutCatalog.definitions.filter {
-                        $0.action.siblingIndex != nil
-                    },
-                    id: \.action
-                ) { definition in
-                    Button(definition.title) {
-                        invoke(definition.action)
-                    }
-                    .keyboardShortcut(shortcut(
-                        definition.action,
-                        actionIsAvailable: availableSiblingShortcuts?
-                            .contains(definition.action) == true
-                    ))
-                    .disabled(
-                        availableSiblingShortcuts?.contains(
-                            definition.action
-                        ) != true
-                    )
-                }
-            }
-
-            Divider()
-
-            Button("Apply Theme to Current Session") {
-                NotificationCenter.default.post(
-                    name: .ghosthubApplyThemeToCurrentSession,
-                    object: nil
-                )
-            }
-            .disabled(
-                focusedSceneModel?
-                    .canApplyThemeToActiveTmuxSession != true
-            )
-        }
-    }
-
-    // MARK: - View menu
-
-    @CommandsBuilder
-    private var viewMenuCommands: some Commands {
-        CommandGroup(after: .sidebar) {
-            Divider()
-            Button("Command Palette") {
-                invoke(.commandPalette)
-            }
-            .keyboardShortcut(shortcut(.commandPalette))
-            .disabled(focusedSceneModel == nil)
-
-            Divider()
-
-            Button("Toggle Sidebar") {
-                invoke(.toggleSidebar)
-            }
-            .keyboardShortcut(shortcut(.toggleSidebar))
-            .disabled(focusedSceneModel == nil)
-        }
-    }
-
-    private func shortcut(
-        _ action: ApplicationShortcutAction,
-        actionIsAvailable: Bool = true
-    ) -> KeyboardShortcut? {
-        ApplicationShortcutMenuModel.keyboardBinding(
-            settingsStore.shortcutPreferences.resolved[action],
-            sceneIsFocused: focusedSceneModel?.isFocusedWindow == true,
-            hasAttachedSheet: focusedSceneHasAttachedSheet,
-            actionIsAvailable: actionIsAvailable
-        )?.swiftUI
-    }
-
-    private func splitShortcut(
-        _ action: ApplicationShortcutAction
-    ) -> KeyboardShortcut? {
-        let splitBinding = ApplicationShortcutMenuModel.splitBinding(
-            settingsStore.shortcutPreferences.resolved[action],
-            terminalHasEffectiveKeyboardFocus:
-            terminalHasEffectiveKeyboardFocus
-        )
-        return ApplicationShortcutMenuModel.keyboardBinding(
-            splitBinding,
-            sceneIsFocused: focusedSceneModel?.isFocusedWindow == true,
-            hasAttachedSheet: focusedSceneHasAttachedSheet,
-            actionIsAvailable: focusedSceneModel?.canSplitActivePane == true
-        )?.swiftUI
-    }
-
-    private var focusedSceneHasAttachedSheet: Bool {
-        guard let focusedSceneModel else { return false }
-        return focusedSceneModel.isSettingsPresented
-            || focusedSceneModel.isCommandPalettePresented
-            || focusedSceneModel.isLogViewerPresented
-    }
-
-    private func invoke(_ action: ApplicationShortcutAction) {
-        let invocation = ApplicationShortcutMenuModel.invocation(
-            currentEvent: NSApplication.shared.currentEvent,
-            binding: settingsStore.shortcutPreferences.resolved[action]
-        )
-        _ = focusedSceneModel?.performApplicationShortcut(
-            action,
-            invocation: invocation
-        )
-    }
-}
-
-extension ApplicationKeyBinding {
-    var swiftUI: KeyboardShortcut {
-        KeyboardShortcut(swiftUIKey, modifiers: swiftUIModifiers)
-    }
-
-    var swiftUIKey: KeyEquivalent {
-        switch key {
-        case .character("+"): KeyEquivalent("=")
-        case let .character(character): KeyEquivalent(character)
-        case .tab: .tab
-        case .return: .return
-        case .escape: .escape
-        case .delete: .delete
-        case .leftArrow: .leftArrow
-        case .rightArrow: .rightArrow
-        case .upArrow: .upArrow
-        case .downArrow: .downArrow
-        case let .function(number):
-            KeyEquivalent(Character(UnicodeScalar(0xF703 + number)!))
-        }
-    }
-
-    var swiftUIModifiers: EventModifiers {
-        var result: EventModifiers = []
-        if modifiers.contains(.command) {
-            result.insert(.command)
-        }
-        if modifiers.contains(.control) {
-            result.insert(.control)
-        }
-        if modifiers.contains(.option) {
-            result.insert(.option)
-        }
-        if modifiers.contains(.shift) {
-            result.insert(.shift)
-        }
-        if key == .character("+") {
-            result.insert(.shift)
-        }
-        return result
-    }
 }
